requirements.txt
blinker==1.9.0
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
defusedxml==0.7.1
Flask==3.1.0
google-api-core==2.24.0
google-api-python-client==2.156.0
google-auth==2.37.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.66.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
proto-plus==1.25.0
protobuf==5.29.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.0
python-dotenv==1.0.1
requests==2.32.3
rsa==4.9
uritemplate==4.1.1
urllib3==2.2.3
Werkzeug==3.1.3
youtube-transcript-api==0.6.3
altair==4.2.0
attrs==22.1.0
beautifulsoup4==4.11.1
blinker==1.5
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
clean-text==0.6.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cssselect==1.1.0
decorator==5.1.1
emoji==1.7.0
entrypoints==0.4
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
ftfy==6.1.1
gitdb==4.0.9
GitPython==3.1.27
idna==3.4
importlib-metadata==5.0.0
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
lxml>=4.9.0
lxml_html_clean>=0.1.0
MarkupSafe==2.1.1
newspaper3k==0.2.8
nltk==3.7
numpy==1.23.3
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
protobuf==3.20.3
pyarrow==9.0.0
pydeck==0.8.0b3
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2022.9.13
requests==2.28.1
requests-file==1.5.1
rich==12.6.0
semver==2.13.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
streamlit==1.13.0
textblob==0.17.1
tinysegmenter==0.3
tldextract==3.3.1
toml==0.10.2
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
typing_extensions==4.3.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
zipp==3.8.1